% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddPCR_uniplex.r
\name{ddPCR_uniplex}
\alias{ddPCR_uniplex}
\title{Calculate the number of target copies from a uniplex ddPCR output}
\usage{
ddPCR_uniplex(n_positive, n_droplets)
}
\arguments{
\item{n_positive}{Number of positive droplets among \code{n_total} (integer,
possibly a vector representing multiple measurements).}

\item{n_droplets}{Total number of droplets (integer vector of the same length
as \code{n_positive}).}
}
\value{
The estimated number of target copies (numeric).
}
\description{
Given the number of positive droplets and the total number of droplets,
the function computes an estimate of the number of copies in the sample.
If the percentage of positives is very low such that all droplets contain
a single copy of the target only, this is trivial. However, at a higher
percentage of positives, we need to account for the fact that some droplets
will contain more than one copy. Hence, we need to use the Poisson
distribution.
}
\note{
The plain solution is

   \deqn{copies = -1 * ln(1 - n_positive / n_droplets) * n_droplets}

with \eqn{ln} being the natural logarithm. See comments in the code for the
derivation.
}
\examples{
n_droplets <- 1000
n_positive <- seq(from=1, to=200, by=10)
copies <- sapply(n_positive, ddPCR_uniplex, n_droplets=n_droplets)
plot(n_positive, copies, xlab="Positive droplets", ylab="Copies in sample")
legend("bottomright", bty="n", lty=c(1,NA), pch=c(NA,1),
  legend=c("1:1 relation","estimates based Poisson distribution"))
abline(0, 1)
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
