% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_read.r
\name{fasta_read}
\alias{fasta_read}
\title{Read sequences in FASTA format}
\usage{
fasta_read(f)
}
\arguments{
\item{f}{File path as a character string.}
}
\value{
A named vector of type character, each element of which holds as
  sequence from the input file. Element names correspond to the sequence
   identifiers after removal of the initial ">" character.
}
\description{
Reads sequences from a file in FASTA format. The function is applicable to
files with fixed-width formatting (sequences wrapped by newlines) as well as
to files without line breaks (single line per sequence).
}
\note{
The function may fail on very large input files (e.g. metagenomes)
  not fitting into memory as a whole. Typical cases of file corruption
  are recognized and an error is thrown.
}
\examples{
# create sample file with and without line breaks
f <- tempfile()
write(c(">seq 1", "ATGCCTA", ">seq2", "AGATA", "CAT"), file=f)

# import sequences
s <- fasta_read(f)

rm(f)
print(s)
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
