% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_write.r
\name{fasta_write}
\alias{fasta_write}
\title{Write sequences in FASTA format}
\usage{
fasta_write(x, f)
}
\arguments{
\item{x}{A named vector of type character holding sequence information.}

\item{f}{File path as a character string.}
}
\description{
Exports a vector of sequences in FASTA format.
}
\note{
Sequences are written to the file without line breaks. Sequence headers
  are constructed from element names by appending the usual ">" as a suffix.
}
\examples{
# export sequence
f <- tempfile()
fasta_write(c(`seq 1`="ATG", `seq2`="CAT"), f)

# re-import
s <- fasta_read(f)

rm(f)
print(s)
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
