% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qPCR.R
\name{qPCR}
\alias{qPCR}
\title{Function to automate the analysis of qPCR data}
\usage{
qPCR(data, target, reference, calibrator, rm_files = "none", plot = FALSE)
}
\arguments{
\item{data}{data.frame of tabular qPCR data}

\item{calibrator}{character with name of sample used for calibration}

\item{rm_files}{character vector containing names of samples to remove}

\item{plot}{logical indicating whether to plot or return original results with added
"ratio" column, providing relative fold changes in expression, relative to calibrator.}

\item{target_gene}{character with name of target gene used in analysis}

\item{ref_gene}{character with name of reference gene used in analysis}
}
\description{
Designed for tabular qPCR data of a specific format generated by the Ernst Lab at
 Michigan State University. To see an example of the data, load an example external dataset with
 data("adar_data").
}

